% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agd_2_csv.R
\name{agd_2_csv}
\alias{agd_2_csv}
\title{Convert agd Files to CSV}
\usage{
agd_2_csv(
  agd_files,
  outdir = NULL,
  vm = TRUE,
  progress = FALSE,
  parallel = TRUE,
  cores = NULL,
  logfile = FALSE,
  verbose = FALSE,
  recursive = TRUE
)
}
\arguments{
\item{agd_files}{The file or files to convert. You can provide any of a
path to a single file, a path to a directory, or a vector of file paths.}

\item{outdir}{A directory where converted CSV files will be saved. If NULL,
the files are saved in the same directory as the original files.}

\item{vm}{Should the vector magnitude column be included? Defaults to TRUE.}

\item{progress}{Display a progress bar. Defaults to TRUE.}

\item{parallel}{Use a parallel backend. Defaults to FALSE.}

\item{cores}{If \code{parallel == TRUE}, how many cores are used for processing.
By default this is the smaller of the number of cores available minus 1, or
the number of files to process.}

\item{logfile}{create a log file for debugging. Can be one of \code{FALSE}
(default; do not create a file), \code{TRUE} (create the log file at the default
location), or a path for where to store the log file.}

\item{verbose}{Logical for if additional information should be displayed.
Defaults to \code{FALSE}}

\item{recursive}{If agd_files is a directory, should
sub-folders be searched for agd files?}
}
\value{
Nothing. Files are saved as a side effect.
}
\description{
Convert a agd file to the CSV format provided by ActiLife.
}
\examples{
\dontrun{
agd_2_csv(
  agd_files = my_directory,
  outdir = NULL, # Save to the same place
  progress = FALSE, # Show a progress bar?
  parallel = TRUE # Process files in parallel?
)
}
}
