% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_input}
\alias{check_file_input}
\title{Check the Input Type for the Main Function}
\usage{
check_file_input(accel_files)
}
\arguments{
\item{accel_files}{The parameter to be checked.}
}
\value{
One of "single", "directory", or "vector".
}
\description{
Not intended to be called directly. Checks if the gt3x_files or agd_files
parameter is a single file, a directory, or a vector of file paths. Errors
if no files are found.
}
