% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agd_2_csv.R
\name{convert_agd_file}
\alias{convert_agd_file}
\title{Convert a Single agd File}
\usage{
convert_agd_file(
  agd_file,
  outfile,
  cores,
  dateformat = "\%Y-\%m-\%d",
  include_timestamp = FALSE
)
}
\arguments{
\item{agd_file}{A path to a single agd file}

\item{outfile}{The path to save the resulting CSV file}

\item{cores}{The number of cores to use for parallel processing.}

\item{dateformat}{The format to use for the date string in the header.}

\item{include_timestamp}{Should the timestamp column be included in the
output?}
}
\value{
nothing
}
\description{
Not intended to be called directly. The function to drive the foreach loop.
}
