% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{convert_file}
\alias{convert_file}
\title{Convert a Single GT3X File}
\usage{
convert_file(gt3x_file, outfile, actilife = FALSE)
}
\arguments{
\item{gt3x_file}{A path to a single GT3X file}

\item{outfile}{The path to save the resulting CSV file}

\item{actilife}{The version string for the header. By default, this is
"gt3x2csv v0.2.0". If your analysis depends on knowing an Actilife version,
you can provide one here.}
}
\value{
nothing
}
\description{
Not intended to be called directly. The function to drive the foreach loop.
}
