% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divide_1e7}
\alias{divide_1e7}
\title{divide_1e7}
\usage{
divide_1e7(y)
}
\arguments{
\item{y}{value to be divided by 1e7}
}
\value{
double number z = y/1e7
}
\description{
util function to divide the date-time number provided by the .gt3x file by 1e7, in order to make possible the conversion to dd/mm/yyyy
}
