% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_outputfiles}
\alias{generate_outputfiles}
\title{Create File Names for Output Files}
\usage{
generate_outputfiles(gt3x_files, outdir = NULL)
}
\arguments{
\item{gt3x_files}{The file or files to for which output locations will be
made.}

\item{outdir}{Optionally provide a single directory for all CSV files to be
stored in.}
}
\value{
A vector of file paths.
}
\description{
Creates the vector of file paths that the resulting CSV files will be saved
to. Not intended to be called by users.
}
