% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x2csv.R
\docType{package}
\name{gt3x2csv}
\alias{gt3x2csv}
\alias{gt3x2csv-package}
\title{gt3x2csv: A package to convert Actigraph proprietary .gt3x files to csv format.}
\description{
: This project was designed with the intent of converting files from the gt3x Actigraph generated files to csv files with the same structure used when exporting/converting raw accelerometry data in the csv format directly from the actilife software.
}
\section{Foo functions}{

gt3x_2_csv, gt3x_folder_2_csv, gt3x_2_csv_par
}

