% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{gt3x_2_csv}
\alias{gt3x_2_csv}
\title{Convert GT3X Files to Raw CSV}
\usage{
gt3x_2_csv(
  gt3x_files,
  outdir = NULL,
  progress = FALSE,
  parallel = TRUE,
  cores = parallel::detectCores() - 1,
  logfile = FALSE,
  verbose = FALSE,
  recursive = TRUE
)
}
\arguments{
\item{gt3x_files}{the file or files to convert. You can provide any of a
path to a single file, a path to a directory, or a vector of file paths.}

\item{outdir}{a directory where converted CSV files will be saved. If NULL,
the files are saved in the same directory as the original files. If
gt3x_files is a directory and \code{recursive==TRUE}, the folder structure is
replicated in outdir.}

\item{progress}{Display a progress bar. Defaults to TRUE.}

\item{parallel}{Use a parallel backend. Defaults to TRUE.}

\item{cores}{If \code{parallel == TRUE}, how many cores are used for processing.}

\item{logfile}{create a log file for debugging. Can be one of \code{FALSE}
(default; do not create a file), \code{TRUE} (create the log file at the default
location), or a path for where to store the log file.}

\item{verbose}{logical for if additional information should be displayed.
Defaults to \code{FALSE}}

\item{recursive}{if gt3x_files is a directory, should sub-folders be searched
for GT3X  files?}
}
\value{

}
\description{
TODO ADD DESCRIPTION
}
