% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv_par.R
\name{gt3x_2_csv_par}
\alias{gt3x_2_csv_par}
\title{gt3x_2_csv_par}
\usage{
gt3x_2_csv_par(folder, n.cores = detectCores() - 1)
}
\arguments{
\item{folder}{The folder where the files are located}

\item{n.cores}{Default = ncores()-1. Specify the number of cores to be used in the operation}
}
\value{
a folder named csv with inside the folder where the .gt3x files are located
}
\description{
Processing multiple gt3x files at the same time
}
\details{
Processes all the files of a given folder in pararllel using foreach function
}
\seealso{
gt3x_2_csv

gt3x_folder_2_csv
}
