% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_folder_2_csv.R
\name{gt3x_folder_2_csv}
\alias{gt3x_folder_2_csv}
\title{gt3x_folder_2_csv}
\usage{
gt3x_folder_2_csv(folder_path)
}
\arguments{
\item{folder_path}{The folder where the files are located}
}
\value{
A folder named "csv" inside the folder where the .gt3x files are located
}
\description{
Converts al the files inside a folder
}
\details{
Lists all the .gt3x files in a given folder and converts it to the actilife RAW data csv file
}
\seealso{
gt3x_2_csv

gt3x_2_csv_par
}
