% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{header_csv}
\alias{header_csv}
\title{GT3X header to Actilife's format}
\usage{
header_csv(origin, outdir, verbose = FALSE)
}
\arguments{
\item{origin}{path to a GT3X file to convert}

\item{outdir}{path for outputting the CSV formatted file}

\item{verbose}{logical: wether to show detailed log messages or not (default: FALSE)}
}
\description{
Reads metadata from ``.gt3x`` files, formats it to Actilife's standard format (for compatibility with GGIR) and exports it to a CSV file.
}
\details{
Reads the metadata from a GT3X file's internal "info.txt" file using the read_info function, then exports it as the header to a CSV file using the save_header function.
}
