% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_gt3x_rec}
\alias{list_gt3x_rec}
\title{Convert a Directory to a Vector of Files}
\usage{
list_gt3x_rec(path, recursive = TRUE)
}
\arguments{
\item{path}{The path the search for files.}

\item{recursive}{Should sub-folders also be searched? Defaults to TRUE.}
}
\value{
A vector of file paths.
}
\description{
Searches a directory for gt3x files. Not intended to be called by users.
}
