% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{read_info}
\alias{read_info}
\title{Read metadata from a GT3X file}
\usage{
read_info(origin, verbose = FALSE)
}
\arguments{
\item{origin}{path to a GT3X file from which to read the metadata}

\item{verbose}{logical: wether to show detailed log messages or not (default: FALSE)}
}
\description{
Read the metadata from a GT3X file as a data.frame.
}
\details{
Given a complete path to a file with extension ".gt3x" (exported from Actilife's software), import it's internal "info.txt" file, extract it's metadata and format it into an R data.frame.
}
