% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round2}
\alias{round2}
\title{Alternative Rounding}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{The number to round (or vector, or data.frame)}

\item{n}{The number of digits to round to.}
}
\value{
A vector of rounded numbers
}
\description{
Alternative Rounding
}
\examples{
# Examples of the standard `round` function
round(0.3)
round(0.5)
round(0.7)
round(0.15, 1)

# Examples of rounding that rounds `0.5` up
round2(0.3)
round2(0.5)
round2(0.7)
round2(0.15, 1)
}
