% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agd_2_csv.R
\name{save_agd_data}
\alias{save_agd_data}
\title{Write Activity Data to CSV}
\usage{
save_agd_data(con, outfile, cores, include_timestamp)
}
\arguments{
\item{con}{The connection to the agd file established by \code{DBI::dbConnect()}}

\item{outfile}{The path to save the resulting CSV file}

\item{cores}{The number of cores allocated to parallel processing}

\item{include_timestamp}{Should the timestamp column be included in the
output?}
}
\value{
Nothing
}
\description{
Write the activity data to a CSV file. Not intended to be called directly.
}
