% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agd_2_csv.R
\name{save_agd_header}
\alias{save_agd_header}
\title{Save Information to CSV Header}
\usage{
save_agd_header(con, outfile, dateformat)
}
\arguments{
\item{con}{The connection to the agd file established by \code{DBI::dbConnect()}}

\item{outfile}{The path to save the resulting CSV file}

\item{dateformat}{The format to use for the date string in the header.}
}
\value{
nothing
}
\description{
Reads information from agd file and writes it in the expected format to a CSV
file. Not intended to be called directly.
}
