% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_header}
\alias{save_header}
\title{save_header}
\usage{
save_header(df_file = info_filedf, dest_csv = csv_folder, file_id)
}
\arguments{
\item{dest_csv}{default = destination folder  the folder to which you want to generate the header file}

\item{infofile}{default = info_filedf (output from the [read_info] function) data frame containing the metadata generated through the read_info function}
}
\value{
a .csv file named "identificationRAW.csv" wuth the header
}
\description{
Saves .gt3x metadata as csv header
}
\details{
Saves the header extracted from the .gt3x file with the read_info function in the .csv extension (look at read_info function)
}
