% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_log}
\alias{setup_log}
\title{Setup the Logging for Data Processing}
\usage{
setup_log(logfile, verbose, outdir)
}
\arguments{
\item{logfile}{create a log file for debugging. Can be one of \code{FALSE}
(default; do not create a file), \code{TRUE} (create the log file at the default
location), or a path for where to store the log file.}

\item{verbose}{Logical for if additional information should be displayed.
Defaults to \code{FALSE}.}

\item{outdir}{A directory the logfile will be saved}
}
\value{
Nothing.
}
\description{
Setup the Logging for Data Processing
}
