% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_dates}
\alias{transform_dates}
\title{transform_dates}
\usage{
transform_dates(x)
}
\arguments{
\item{x}{default = output of the [divide_1e7] function. the object containing the date to be changed.}
}
\value{
the date-ime from the .txt file in the "dd/mm/yyyy" format
}
\description{
Converting the date-time information to the dd-mm-yy format
}
\details{
Util to change the format of the date-times to the format used in the .csv file. The date-time of the txt file come in the format: number of days from 0001-01-01.
}
