% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_gt3x_files}
\alias{validate_gt3x_files}
\title{Check the GT3X files are Valid}
\usage{
validate_gt3x_files(gt3x_files, proc_type)
}
\arguments{
\item{gt3x_files}{The file or files to check. Either a single file or a
vector of file paths.}

\item{proc_type}{The format of gt3x_files. Either "single", "directory", or
"vector".}
}
\value{
Nothing. Errors if invalid files are found.
}
\description{
Validates GT3X files using \code{read.gt3x::have_log_and_info}. Not intended to be
called by users.
}
