% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{gt3x_2_csv}
\alias{gt3x_2_csv}
\title{gt3x_2_csv}
\usage{
gt3x_2_csv(gt3x_file)
}
\arguments{
\item{path}{the path to the given file}

\item{dest_csv}{desired destination folder}
}
\description{
Converts a given .gt3x file to .csv format
}
\details{
Reads both the .txt file and the .bin file located inside the .gt3x file given by actilife software and converts it to a csv file in the save format of the .csv file extracted from the sofrtware.
}
\seealso{
gt3x_folder_2_csv converts a folder

gt3x_2_csv_par converts a a folder using paralell processing
}
