% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{gt3x_2_csv}
\alias{gt3x_2_csv}
\title{Export GT3X as CSV}
\usage{
gt3x_2_csv(
  gt3x_files = NULL,
  outdir = NULL,
  progress = FALSE,
  parallel = FALSE,
  cores = detectCores(),
  logfile = "default",
  verbose = FALSE
)
}
\arguments{
\item{gt3x_files}{complete path to a GT3X file to convert. Can also be a vector of complete filenames or a directory with ".gt3x" files in it.}

\item{outdir}{path to a directory for outputting the CSV formatted file (default: NULL). There will be another directory created inside this one, called "/csv".}

\item{progress}{logical: wether or not to show TCLTK progress bar (default: FALSE)}

\item{parallel}{logical: wether or not to use parallel processing (default: FALSE)}

\item{cores}{integer: number of cores to use for parallelization. If parallel = FALSE, this argument will be ignored.}

\item{logfile}{path to use for outputting logfile. If FALSE, even when verbose = TRUE only print basic info messages and directly to R log. Default ('default') outputs to R's log when parallel = FALSE and to the same directory as the CSV output when parallel = TRUE.}

\item{verbose}{logical: wether or not to show detailed log messages (default: FALSE)}
}
\description{
Convert GT3X files to CSV format in Actilife's default export format.
}
\details{
Reads the "info.txt" and "log.bin" internal files from ".gt3x" files exported by Actilife's software and converts it to a CSV file in the same format of the CSV file extracted directly from the software.
}
