% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{header_csv}
\alias{header_csv}
\title{header_csv}
\usage{
header_csv(origin)
}
\arguments{
\item{origin}{the path to the .gt3xfile to be converted}
}
\description{
Reads the metadata from gt3x files and savaes as csv
}
\details{
Reads the metadata from the txt file located inside the .gt3x file provided by actigraph using the read_info function and saves it as a csv document using the save_header function.
}
