% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{read_info}
\alias{read_info}
\title{Read Info}
\usage{
read_info(file_txt = file_txt)
}
\arguments{
\item{file_txt}{The path to the desired .txt file}
}
\description{
Reads the metadata of the gt3x file
}
\details{
Reads the metadata registered in the .txt file that is contained inside .gt3x file provided by the actilife software
}
