% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{save_accel}
\alias{save_accel}
\title{Write Activity Data to CSV}
\usage{
save_accel(gt3x_file, outfile)
}
\arguments{
\item{gt3x_file}{An object read using \code{read.gt3x::read.gt3x()}}

\item{outfile}{The path to save the resulting CSV file}
}
\value{
Nothing
}
\description{
Write the activity data to a CSV file. Not intended to be called directly.
}
