% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{save_accel}
\alias{save_accel}
\title{save_accel}
\usage{
save_accel(acc.file)
}
\arguments{
\item{acc.file}{the path to te .gt3x file}
}
\description{
Saves acceleration of the given file
}
\details{
Reads the binary data inside the .gt3x file and saves it in .csv format
}
