% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{save_header}
\alias{save_header}
\title{Save Information to CSV Header}
\usage{
save_header(gt3x_file, outfile, actilife = FALSE)
}
\arguments{
\item{gt3x_file}{An object read using \code{read.gt3x::read.gt3x()}}

\item{outfile}{The path to save the resulting CSV file}

\item{actilife}{The version string for the header. By default, this is
"gt3x2csv v0.2.0". If your analysis depends on knowing an Actilife version,
you can provide one here.}
}
\value{
nothing
}
\description{
Reads information from the 'info.txt' file and writes it in the expected
format to a CSV file. Not intended to be called directly.
}
