% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{save_header}
\alias{save_header}
\title{save_header}
\usage{
save_header(df_file = info_filef, dest_csv = csv_folder, file_id)
}
\arguments{
\item{dest_csv}{default = ddestination folder. the folder to which you want to generate the header file}

\item{infofile}{default = info_filef data frame containing the metadata generated through the read_info function}

\item{files_list_i}{the name of the file that is going to be saved}
}
\description{
Saves .gt3x metadata as header
}
\details{
Saves the header extracted from the .gt3x file with the read_info function in the .csv extension (look at read_info function)
}
