% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{save_header}
\alias{save_header}
\title{Export GT3X metadata to CSV}
\usage{
save_header(df_file, outfile_csv, verbose = FALSE)
}
\arguments{
\item{df_file}{data.frame with metadata generated by the read_info function}

\item{outfile_csv}{complete path to the output CSV file the header will be written to}

\item{verbose}{logical: wether to show detailed log messages or not (default: FALSE)}
}
\description{
Formats the data.frame obtained from the read_info function into Actilife's header format, then exports it to a CSV file.
}
\details{
Saves the header extracted from the .gt3x file with the read_info function in the .csv extension (look at read_info function)
}
