% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{substrRight}
\alias{substrRight}
\title{substrRight}
\usage{
substrRight(x, n = nchar(x) - j, j = 0)
}
\arguments{
\item{x}{the string to be manipulated}

\item{n}{the number of characters to extract from the string}

\item{j}{the number of characters to ignore in the end of the string}
}
\description{
Extracting n characters from the end of the string "jumping" last j characters
}
