% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_2_csv.R
\name{transform_dates}
\alias{transform_dates}
\title{transform_dates}
\usage{
transform_dates(x)
}
\arguments{
\item{x}{the object containing the date to be changed.}
}
\description{
Changing the formatting of the date-time information
}
\details{
gt3x_2_csv

Util to change the format of the date-times to the format used in the .gt3x file
}
